Clarify home page value proposition
